requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
httpx==0.27.2
python-multipart==0.0.12
passlib==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
twilio==9.3.3
alembic==1.13.3
psycopg2-binary==2.9.9

